requirements.txt
python-dotenv
aiogram==3.4.1
fastapi
uvicorn